♻️ refactor: 리크루팅 홈 폴더 구조 수정정
